pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "startupscout"
version = "1.0.0"
description = "Production-grade web data extraction platform for startup intelligence"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}

dependencies = [
    "httpx[http2]>=0.27.0",
    "playwright>=1.40.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "structlog>=24.1.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.1.0,<6.0.0",
    "groq>=0.4.0",
    "python-Levenshtein>=0.25.0",
    "tqdm>=4.66.0",
    "python-dotenv>=1.0.0",
    "pandas>=2.1.0",
    "tenacity>=8.2.0",
    "fake-useragent>=1.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.2.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
